refactor our root store into slices and use logging actions names for easier debugging
