+++ src/store/index.ts
@@ -1,5 +1,17 @@
-import useTreeStore from 'store/treeStore';
+import { createTreeSlice, TreeSlice } from 'store/treeSlice';
+import { create } from 'zustand';
+import { devtools } from 'zustand/middleware';
 
-export type { DecisionTree, TreeNode, TreeStore, DagNode } from './treeStore';
-export { useTreeStore };
+export type { DecisionTree, TreeNode, TreeSlice, DagNode } from 'store/treeSlice';
 export { buildTreeEdges } from 'store/treeUtils';
+
+export const useStore = create<TreeSlice>()(
+  devtools(
+    (...args) => ({
+      ...createTreeSlice(...args),
+    }),
+    { name: 'The Manifest Game', store: 'decisionTree' }
+  )
+);
+
+export default useStore;
